backend/pyproject.toml
[tool.poetry]
name = "nvd_backend"
version = "1.0.0"
description = "API + Sync server for the Arduino"
authors = ["Roshan Pawar <steradian614@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
fastapi = "0.68.1"
pyserial = "3.5"
opencv-python = "4.5.5.64"
Pillow = "9.0.1"
loguru = "0.6.0"

[tool.poetry.dev-dependencies]
black = "*"
autoflake = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
